bug(ui.mask): change required check in parser
Change parsers check for required error to an attribute check.

Before the parser checked to see if the controller.$error.required was defined.
If so, it would set the required validity to false. This doesn't work if
ng-required is used, because the control could have been required at one point
and is no longer required. Instead this checks if the elements attribute object
contains a required property and if said property is true.

BREAKING CHANGE: if you rely on ui-mask to invalidate an empty viewValue
this will no longer work. Please use the required attribute or ng-required
directive to specify whether the input is required.

fixes #198